+++ generators/create:valueobject/templates/type/data.ts
@@ -1,0 +1,13 @@
+<%= typescriptHeader %>
+
+/**
+ * `<%= typeName %>`Data is the input required to make a new
+ * {@link <%= typeName %>}
+ *
+ * @category <%= typeName %>
+ */
+export interface <%= typeName %>Data {
+    // @TODO
+    //
+    // add the details here!
+}